move retry queries to dashboard, drop unused functions
