+++ src/Beckett.Dashboard/Subscriptions/Checkpoints/BulkRetry/BulkRetryEndpoint.cs
@@ -1,4 +1,5 @@
-using Beckett.Subscriptions.Retries;
+using Beckett.Dashboard.Subscriptions.Checkpoints.Shared.Queries;
+using Beckett.Database;
 
 namespace Beckett.Dashboard.Subscriptions.Checkpoints.BulkRetry;
 
@@ -7,11 +8,12 @@ public static class BulkRetryEndpoint
     public static async Task<IResult> Handle(
         HttpContext context,
         [FromForm(Name = "id")] long[] ids,
-        IRetryClient retryClient,
+        IPostgresDatabase database,
+        PostgresOptions options,
         CancellationToken cancellationToken
     )
     {
-        await retryClient.BulkRetry(ids, cancellationToken);
+        await database.Execute(new ScheduleCheckpoints(ids, DateTimeOffset.UtcNow, options), cancellationToken);
 
         context.Response.Headers.Append("HX-Refresh", new StringValues("true"));
         context.Response.Headers.Append("HX-Trigger", new StringValues("bulk_retry_requested"));
