+++ src/Beckett.Dashboard/Subscriptions/Checkpoints/Retry/RetryEndpoint.cs
@@ -1,4 +1,5 @@
-using Beckett.Subscriptions.Retries;
+using Beckett.Dashboard.Subscriptions.Checkpoints.Shared.Queries;
+using Beckett.Database;
 
 namespace Beckett.Dashboard.Subscriptions.Checkpoints.Retry;
 
@@ -7,11 +8,12 @@ public static class RetryEndpoint
     public static async Task<IResult> Handle(
         HttpContext context,
         long id,
-        IRetryClient retryClient,
+        IPostgresDatabase database,
+        PostgresOptions options,
         CancellationToken cancellationToken
     )
     {
-        await retryClient.Retry(id, cancellationToken);
+        await database.Execute(new ScheduleCheckpoints([id], DateTimeOffset.UtcNow, options), cancellationToken);
 
         context.Response.Headers.Append("HX-Refresh", new StringValues("true"));
         context.Response.Headers.Append("HX-Trigger", new StringValues("retry_requested"));
