+++ src/Beckett.Dashboard/Subscriptions/Checkpoints/Skip/SkipEndpoint.cs
@@ -1,4 +1,4 @@
-using Beckett.Subscriptions.Retries;
+using Beckett.Database;
 
 namespace Beckett.Dashboard.Subscriptions.Checkpoints.Skip;
 
@@ -7,11 +7,12 @@ public static class SkipEndpoint
     public static async Task<IResult> Handle(
         HttpContext context,
         long id,
-        IRetryClient retryClient,
+        IPostgresDatabase database,
+        PostgresOptions options,
         CancellationToken cancellationToken
     )
     {
-        await retryClient.Skip(id, cancellationToken);
+        await database.Execute(new SkipQuery(id, options), cancellationToken);
 
         context.Response.Headers.Append("HX-Refresh", new StringValues("true"));
         context.Response.Headers.Append("HX-Trigger", new StringValues("checkpoint_skipped"));
